Add libCZIAPI (#134)
* add libcziapi-code

* merge changes to libCZI - this should build

* change default of "LIBCZI_BUILD_LIBCZIAPI" (to off)

* test

* test

* test

* test

* test

* move Doxyfile

* Update version history and add resolution protocol

- Updated `version_history.md` to include versions 0.63.0 to 0.64.0, highlighting new features and fixes.
- Added a new section in `resolution_protocol.markdown` detailing guidelines for handling discrepancies in CZI files, focusing on consistent behavior for malformed data and specific recommendations for resolving conflicts.

* test

* Update resolution protocol section heading

Changed the heading for the "resolution protocol" section to a higher-level format. This change emphasizes the significance of the topic and includes a brief description of its purpose, which addresses ambiguous or contradictory information.

* remove now obsolete documentation-files

* Update GitHub Pages deployment workflow

Added a new "Pages" workflow in `pages.yml` that triggers on `workflow_dispatch` and `pull_request` events. Removed the previous deployment workflow from `pages_orig.y_m_l`, which included jobs for Doxygen documentation. The new configuration aims to streamline the deployment process.

* Add documentation for libCZI_WriterClose function

Enhanced the documentation comments for the `libCZI_WriterClose` function. The comments now explain the function's purpose, emphasize the importance of calling it for proper file finalization, and provide details on the parameters and return values.

* Update version history for 0.63.1 release

Added entry for version 0.63.1, including a fix for the
CZICmd command "ExtractAttachment" and improvements to
UTF8 handling on Windows.

* remove accidentally checked-in binary

* fix doxygen

* Enhance API documentation and add new functions

* Add coordinate transformation methods in libCZI

Introduce `TransformPoint` and `TransformRectangle` methods in `libCZI.h` for transforming points and rectangles between coordinate systems, with detailed documentation. Update comments in `libCZI_Pixels.h` to clarify the `IntPoint` struct's `x` and `y` coordinates.

* Update IMAGE_PATH in Doxyfile for documentation images

Changed IMAGE_PATH from './libCZI/Doc/Images' to '../../docs/source/_static/images' to reflect the new location of image files used in the documentation.

* Update documentation for TransformRectangle function

Renamed the parameter `source_point` to `source_rectangle` in the `TransformRectangle` function to better reflect its purpose. Updated comments to clarify that it represents a source rectangle along with its coordinate system.

Fix documentation typos and clarify parameter names

Corrected "refered" to "referred" in libCZI.h and updated
the parameter name in TransformRectangle from source_point
to source_rectangle for better clarity.

In libCZIApi.h, modified the "rating" field documentation
to enclose the rating value in quotes, clarifying its
expected data type as a string.

* reduce sphinx-warnings

* cosmetic

* test

* test

* fix links

* test

* test

* test

* Update gradation operation documentation and examples

Enhanced the "Apply gradiation" operation documentation by adding an image reference for the linear gradation curve and expanding the explanation to include definitions for linear, gamma, and spline curves. Clarified the gradation curve generation in `Compositors::ComposeMultiChannel`, including a specific example and updated alt text for the image. Added mentions of utility functions for creating look-up tables from splines and gamma.

* Update documentation formatting for building libCZI

Refactor documentation to use reStructuredText (reST) syntax for links and images. Correctly format the CMake link and add the image directive for the Visual Studio solution screenshot. Update the vcpkg package manager text to use reST link format.

* test

* test

* test

* Update .gitignore and enhance documentation structures

- Updated `.gitignore` to include additional directories and files.
- Clarified comments in `libCZI_Compositor.h` and `libCZI_Write.h`.
- Modified `Doxyfile` to update input files for documentation.
- Added `AccessorOptionsInterop` and `AddAttachmentInfoInterop` structures.
- Enhanced `AddSubBlockInfoInterop`, `ExternalStreamErrorInfoInterop`, and `WriteMetadataInfoInterop` with additional fields and comments.

* Add deploy input for manual GitHub Pages deployment

Introduce a new `deploy` input parameter in `pages.yml` to allow manual triggering of the deployment process. The workflow now checks if the `deploy` input is set to `true` or if the event is a push to the `main` branch before executing deployment steps, enhancing flexibility in managing static content deployment.

* Update licensing information in dep5 file

Replaced LGPL-3.0-or-later license with MIT license.
Expanded the list of files covered by copyright to include
additional project files and removed the previous specification
for `Src/libCZI/Doc/*`.

* fix build

* test

* fix

* test

* update

* cosmetic

* remove accidently checked-in file

* Refactor CMakeLists and remove utility files

Updated CMakeLists.txt to quote source file paths in
LIBCZIAPISRCFILES and removed `src/utils.cpp` and
`src/utils.h`. Deleted the `Message_Box()` function and
its associated Windows compatibility code from the
project.

* Remove unused header from libCZIApi.cpp

Eliminated the inclusion of "utils.h" in `libCZIApi.cpp`. This change suggests that the functions or definitions from "utils.h" are no longer necessary in this file, potentially indicating a refactor or replacement with other implementations.

* remove commented out code

* review

* documentation update

* megalinter error fix

* megalinter error fix

* commit

* commit

* Update Src/libCZI/libCZI_Write.h

Co-authored-by: H.A <50059772+HakanAcundas@users.noreply.github.com>

* review

* Update .gitignore

Co-authored-by: H.A <50059772+HakanAcundas@users.noreply.github.com>

* typo

---------

Co-authored-by: Acundas, Hakan <hakan.acundas@zeiss.com>
Co-authored-by: Hakan Acundas <hakan.acundas@gmail.com>
Co-authored-by: H.A <50059772+HakanAcundas@users.noreply.github.com>