Add support for Azure SDK and CURL streaming options (preparing for vcpkg features) (#149)
* Add support for Azure SDK and CURL streaming options

Introduce configuration flags `LIBCZI_BUILD_AZURESDK_BASED_STREAM`
and `LIBCZI_BUILD_CURL_BASED_STREAM` to enable optional
dependencies for Azure SDK and CURL. Conditional checks are
added to link the necessary dependencies based on the flags,
enhancing the library's flexibility for different streaming
backends during the build process.

* Bump version to 0.66.6 and update version history

Updated the version number in `CMakeLists.txt` to 0.66.6.
Added a new entry in `version_history.md` for version 0.66.6,
which includes preparations for vcpkg features.
Removed duplicate entry for version 0.66.5.

* Update version history for v0.66.6

Added a new entry for version 0.66.6 to the version history, indicating preparation for vcpkg features. This entry complements the existing one for the same version, which is linked to a different pull request.